+++ packages/build-app-helpers/src/getMpaEntries.ts
@@ -1,19 +1,13 @@
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import getRoutesByAppJson from './getRoutesByAppJson';
+import getEntriesByRoute from './getEntriesByRoute';
+import { IEntry } from './types';
 
 interface IOptions {
   target: string;
   appJsonPath?: string;
   appJsonContent?: string;
-}
-
-interface IEntry {
-  entryPath: string;
-  entryName: string;
-  pageName: string;
-  source?: string;
-  path?: string;
 }
 
 // Get entries when exist app.json
@@ -30,35 +24,9 @@ function getEntriesByJson(api, target, appJsonPath, appJsonContent): IEntry[] {
     context: { rootDir },
   } = api;
   const routes = getRoutesByAppJson(target, { appJsonPath, appJsonContent });
-  return routes.map((route) => {
-    let pageName;
-    let entryName;
-    if (route.name) {
-      entryName = route.name;
-      pageName = route.name;
-    } else {
-      const dir = path.dirname(route.source);
-      pageName = path.parse(dir).name;
-      entryName = pageName.toLocaleLowerCase();
-    }
-    return {
-      // routes.pageSource is absolute path, passed from the plugin-store
-      entryPath: route.pageSource || getPageEntryByAppJson(rootDir, route.source),
-      entryName,
-      pageName,
-      source: route.source,
-      path: route.path
-    };
-  });
-}
-
-function getPageEntryByAppJson(rootDir, source) {
-  const absolutePath = path.resolve(rootDir, 'src', source);
-  const targetExt = ['ts', 'tsx', 'js', 'jsx'].find(ext => fs.existsSync(`${absolutePath}.${ext}`));
-  if (!targetExt) {
-    throw new Error(`Cannot find target file ${absolutePath}.`);
-  }
-  return `${absolutePath}.${targetExt}`;
+  return routes.reduce((prev, curr) => {
+    return [...prev, ...getEntriesByRoute(curr, rootDir)];
+  }, []);
 }
 
 function getEntriesByDir(api: any): IEntry[] {