+++ packages/build-app-helpers/src/index.ts
@@ -1,7 +1,10 @@
 export { default as getMpaEntries } from './getMpaEntries';
 export { default as getRoutesByAppJson } from './getRoutesByAppJson';
+export { default as getEntriesByRoute } from './getEntriesByRoute';
 export { default as formatPath } from './formatPath';
 export { default as validation } from './validation';
 export { default as unionBy } from './unionBy';
 export { default as injectTransformRuntime } from './injectTransformRuntime';
 export { default as checkExportDefaultDeclarationExists } from './checkExportDefaultDeclarationExists';
+export { default as getSassImplementation } from './getSassImplementation';
+export { default as getLessImplementation } from './getLessImplementation';