+++ packages/plugin-store/src/babelPluginReplacePath.ts
@@ -9,55 +9,62 @@ const layoutPathRegExp = /src\/pages\/\w+\/Layout/;
 module.exports = ({ types: t }, { routesPath, alias, tempDir, applyMethod }) => {
   return {
     visitor: {
-      ImportDeclaration(nodePath, state) {
-        const isRoutesFile = routesPath.includes(state.filename);
-        if (isRoutesFile) {
-          const { source, specifiers } = nodePath.node;
-          // issue: https://github.com/ice-lab/icejs/issues/271
-          if (t.isImportDefaultSpecifier(specifiers[0]) && specifiers.length === 1) {
-            if (t.isStringLiteral(source)) {
-              const { value } = source;
-              // 约定式路由：
-              // e.g: import Home from '../src/pages/Home/index.tsx';
-              // e.g: import Index from '../src/pages/index.tsx;
-              // 配置式路由：
-              // default alias: import Home from '@/pages/Home';
-              // custom alias: import Home from '$pages/Home';
-              // relative path: import Home from '../pages/Home'
-              const newValue = formatPagePath({ routesPath: state.filename, tempDir, value, alias, applyMethod });
-              // replace to: import Home from 'ice/pages/Home'
-              if (newValue) {
-                replaceWith(t, nodePath, newValue);
-              }
-            }
-          }
-        }
-      },
-
-      CallExpression(nodePath, state) {
-        const isRoutesFile = routesPath.includes(state.filename);
-        if (isRoutesFile) {
-          if (t.isImport(nodePath.node.callee)) {
-            const args = nodePath.node.arguments;
-            for (let i = 0; i < args.length; i++) {
-              const value = args[i].value;
-              if (typeof value === 'string') {
-                // 约定式路由：
-                // e.g: const Home = lazy(() => import(/* webpackChunkName: 'Home' */ '../src/pages/Home/index.tsx'));
-                // 配置式路由：
-                // default alias: const Home = lazy(() => import('@/pages/Home'));
-                // custom alias: const Home = lazy(() => import('$pages/home));
-                // relative path: const Home = lazy(() => import('../pages/Home'));
-                const newValue = formatPagePath({ routesPath: state.filename, tempDir, value, alias, applyMethod });
-                // replace to: const Home =lazy (() => import('ice/Home/Home'));
-                if (newValue) {
-                  args[i].value = newValue;
+      Program: {
+        enter(programPath, state) {
+          programPath.traverse({
+            ImportDeclaration(nodePath) {
+              const isRoutesFile = routesPath.includes(state.filename);
+              if (isRoutesFile) {
+                const { source, specifiers } = nodePath.node;
+                // issue: https://github.com/ice-lab/icejs/issues/271
+                if (t.isImportDefaultSpecifier(specifiers[0]) && specifiers.length === 1) {
+                  if (t.isStringLiteral(source)) {
+                    const { value } = source;
+                    // 约定式路由：
+                    // e.g: import Home from '../src/pages/Home/index.tsx';
+                    // e.g: import Index from '../src/pages/index.tsx;
+                    // 配置式路由：
+                    // default alias: import Home from '@/pages/Home';
+                    // custom alias: import Home from '$pages/Home';
+                    // relative path: import Home from '../pages/Home'
+                    const newValue = formatPagePath({ routesPath: state.filename, tempDir, value, alias, applyMethod });
+                    // replace to: import Home from 'ice/pages/Home'
+                    if (newValue) {
+                      replaceWith(t, nodePath, newValue);
+                    }
+                  }
                 }
               }
-            }
-          }
+            },
+
+            CallExpression(nodePath) {
+              const isRoutesFile = routesPath.includes(state.filename);
+              if (isRoutesFile) {
+                if (t.isImport(nodePath.node.callee)) {
+                  const args = nodePath.node.arguments;
+                  for (let i = 0; i < args.length; i++) {
+                    const value = args[i].value;
+                    if (typeof value === 'string') {
+                      // 约定式路由：
+                      // e.g: const Home = lazy(() => import(/* webpackChunkName: 'Home' */ '../src/pages/Home/index.tsx'));
+                      // 配置式路由：
+                      // default alias: const Home = lazy(() => import('@/pages/Home'));
+                      // custom alias: const Home = lazy(() => import('$pages/home));
+                      // relative path: const Home = lazy(() => import('../pages/Home'));
+                      const newValue = formatPagePath({ routesPath: state.filename, tempDir, value, alias, applyMethod });
+                      // replace to: const Home =lazy (() => import('ice/Home/Home'));
+                      if (newValue) {
+                        args[i].value = newValue;
+                      }
+                    }
+                  }
+                }
+              }
+            },
+          });
         }
-      },
+      }
+
     },
   };
 };